refactor(notifications): retire local parlay notification shim
The TF-967 compatibility shim (parlay-notification-contract.ts, header
still cites @pluto-khronos/types@3.4.3) predated the published parlay
contract. @pluto-khronos/types@3.6.0 now exports
parlayResultNotificationSchema, so route it through the existing
shared-payload-schemas bridge (prefer the published export, local
fallback) the same way propSettled/dailyProps already are, and delete
the shim.

The shim's kind enum matches the published schema exactly. The shim
additionally required payout/actual_payout for won notifications; the
published schema does not — published wins. Updated the util test to
reflect the loosened contract and swapped in an empty-legs rejection.
